Development

+++ packages/ui-components/src/features/tickets/tickets-page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { useContext, useMemo, useState } from "react";
+import { useContext, useMemo } from "react";
 import { getAPI, TicketResponse } from "@app/api";
 import { Pagination } from "../pagination/pagination";
 import { NewTicketDialog } from "./new-ticket-dialog";
@@ -8,21 +8,17 @@ import { SearchBar } from "./search-bar";
 import { DEFAULT_PAGE_SIZE } from "../constants";
 import { useSyncWithSearchParams } from "../../hooks/use-sync-with-search-params";
 import { ToastContext } from "../../toast";
-import { ConfirmationModal } from "../../ConfirmationModal";
 import { TicketsTable } from "./tickets-table";
 import { useQuery } from "@tanstack/react-query";
+import { useConfirmationModal } from "../../confirmation-modal";
 
 export const TicketsPage = () => {
-  const [modalDeleteTicketOpen, setModalDeleteTicket] = useState<{
-    show: boolean;
-    ticket: TicketResponse | null;
-    confirmButtonDisabled: boolean;
-  }>({
-    show: false,
-    ticket: null,
-    confirmButtonDisabled: false,
-  });
   const { newToast } = useContext(ToastContext);
+  const {
+    setConfirmationModalPending,
+    showConfirmationModal,
+    closeConfirmationModal,
+  } = useConfirmationModal();
   const { searchParams, updateSearchParams } = useSyncWithSearchParams();
   const pageNumber = searchParams.get("pageNumber") ?? 1;
   const pageSize = searchParams.get("pageSize") ?? DEFAULT_PAGE_SIZE;
@@ -63,62 +59,58 @@ export const TicketsPage = () => {
   };
 
   const onClickDeleteTicket = (ticket: TicketResponse) => {
-    setModalDeleteTicket({ ...modalDeleteTicketOpen, show: true, ticket });
+    showConfirmationModal(
+      {
+        title: "Ticket löschen",
+        data: ticket,
+        body: <p>Möchten Sie das Ticket wirklich löschen?</p>,
+      },
+      {
+        onConfirm: onConfirmDeleteTicket,
+      }
+    );
   };
 
-  const onConfirmDeleteTicket = async () => {
-    const ticketId = modalDeleteTicketOpen.ticket?.id;
-    if (ticketId) {
-      setModalDeleteTicket({
-        ...modalDeleteTicketOpen,
-        confirmButtonDisabled: true,
-      });
+  const onConfirmDeleteTicket = async (ticket: unknown) => {
+    if (!ticket) {
+      return;
+    }
 
+    const ticketId = (ticket as TicketResponse).id;
+    if (ticketId) {
       try {
+        setConfirmationModalPending(true);
+
         const result = await getAPI().deleteTicket(ticketId);
         if (result.status === "error") {
           newToast({
             msg: "Ticket konnte nicht gelöscht werden: Es fehlen notwendige Berechtigungen.",
             type: "error",
           });
+
+          setConfirmationModalPending(false);
         } else {
           newToast({
             msg: "Ticket erfolgreich gelöscht.",
             type: "success",
           });
+
+          closeConfirmationModal();
         }
       } catch {
         newToast({
           msg: "Ticket konnte nicht gelöscht werden.",
           type: "error",
         });
-      } finally {
-        closeDeleteTicketModal();
+
+        setConfirmationModalPending(false);
       }
     }
-  };
-
-  const closeDeleteTicketModal = () => {
-    setModalDeleteTicket({
-      show: false,
-      confirmButtonDisabled: false,
-      ticket: null,
-    });
   };
 
   return (
     <>
       <NewTicketDialog />
-
-      <ConfirmationModal
-        show={modalDeleteTicketOpen.show}
-        onConfirm={onConfirmDeleteTicket}
-        onCancel={closeDeleteTicketModal}
-        confirmButtonText="Delete"
-        confirmButtonDisabled={modalDeleteTicketOpen.confirmButtonDisabled}
-        title="Ticket löschen"
-        body={<p>Möchten Sie das Ticket wirklich löschen?</p>}
-      />
 
       <div className="mb-4">
         <SearchBar initialSearchText={searchText} onSearch={searchTickets} />
